+++ sdk/identity/Azure.Identity/src/ManagedIdentityClient.cs
@@ -12,8 +12,7 @@ namespace Azure.Identity
     {
         internal const string MsiUnavailableError = "ManagedIdentityCredential authentication unavailable. No Managed Identity endpoint found.";
 
-        private readonly AsyncLockWithValue<ManagedIdentitySource> _identitySourceAsyncLock = new AsyncLockWithValue<ManagedIdentitySource>();
-        private readonly ManagedIdentityClientOptions _options;
+        private readonly ManagedIdentitySource _identitySource;
 
         protected ManagedIdentityClient()
         {
@@ -26,9 +25,14 @@ namespace Azure.Identity
 
         public ManagedIdentityClient(ManagedIdentityClientOptions options)
         {
-            _options = options;
             ClientId = options.ClientId;
             Pipeline = options.Pipeline;
+
+            _identitySource = AppServiceV2017ManagedIdentitySource.TryCreate(options) ??
+                                                    CloudShellManagedIdentitySource.TryCreate(options) ??
+                                                    AzureArcManagedIdentitySource.TryCreate(options) ??
+                                                    ServiceFabricManagedIdentitySource.TryCreate(options) ??
+                                                    new ImdsManagedIdentitySource(Pipeline, ClientId);
         }
 
         internal CredentialPipeline Pipeline { get; }
@@ -37,33 +41,7 @@ namespace Azure.Identity
 
         public virtual async ValueTask<AccessToken> AuthenticateAsync(bool async, TokenRequestContext context, CancellationToken cancellationToken)
         {
-            ManagedIdentitySource identitySource = await GetManagedIdentitySourceAsync(async, cancellationToken).ConfigureAwait(false);
-
-            // if msi is unavailable or we were unable to determine the type return CredentialUnavailable exception that no endpoint was found
-            if (identitySource == default)
-            {
-                throw new CredentialUnavailableException(MsiUnavailableError);
-            }
-
-            return await identitySource.AuthenticateAsync(async, context, cancellationToken).ConfigureAwait(false);
-        }
-
-        private protected virtual async ValueTask<ManagedIdentitySource> GetManagedIdentitySourceAsync(bool async, CancellationToken cancellationToken)
-        {
-            using var asyncLock = await _identitySourceAsyncLock.GetLockOrValueAsync(async, cancellationToken).ConfigureAwait(false);
-            if (asyncLock.HasValue)
-            {
-                return asyncLock.Value;
-            }
-
-            ManagedIdentitySource identitySource = AppServiceV2017ManagedIdentitySource.TryCreate(_options) ??
-                                                    CloudShellManagedIdentitySource.TryCreate(_options) ??
-                                                    AzureArcManagedIdentitySource.TryCreate(_options) ??
-                                                    ServiceFabricManagedIdentitySource.TryCreate(_options) ??
-                                                    await ImdsManagedIdentitySource.TryCreateAsync(_options, async, cancellationToken).ConfigureAwait(false);
-
-            asyncLock.SetValue(identitySource);
-            return identitySource;
+            return await _identitySource.AuthenticateAsync(async, context, cancellationToken).ConfigureAwait(false);
         }
     }
 }